изменил оформление задач 2.3 и 2.4: 1. Убрал преобразования к int or double при выводе строки (до этого так сделал, чтобы ответ вывелся такой же как и в задании) 2. Изменил наименование переменной withdrawal в задании 2.4. 3. Все поправил согласно замечаниям.
Единственное то, что я не переименовывал класси, оставил "hw_2_3". Возьму на заметку, просто для обучения такое название интуитивно понятное.

+++ src/Module2/HomeWork/hw_2_3.java
@@ -25,7 +25,7 @@ public class hw_2_3 {
         if ((double) balanses[numberOfPerson] < (withdrawal + commissions)) {
             return ownerName + " NO";
         } else {
-            return ownerName + " " + (int) withdrawal + " " + (int) ((double) balanses[numberOfPerson] - (withdrawal + commissions));
+            return ownerName + " " + withdrawal + " " + (balanses[numberOfPerson] - (withdrawal + commissions));
         }
 
     }
